restructuring tests/precision
